refactor: move vk social button auth to separate component
